refactor(KERN-1560): extract OR/AND logic into extract-or.ts
Split extract-where.ts (155 lines) into:
- extract-where.ts (75 lines): entry point + residual conjunction builder
- extract-or.ts (90 lines): tryExtractFilters dispatcher, extractOrFilter, extractOrBranch

work-item: KERN-1560
